refactor(scheduler): scope snapshot-to-test to docs and cmd/snapshot-tool
